backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
django==5.1.1; python_version >= '3.10'
django-ckeditor==6.7.1; python_version >= '3.8'
django-ckeditor-5==0.2.13; python_version >= '3.7'
django-cors-headers==4.4.0; python_version >= '3.8'
django-js-asset==2.2.0; python_version >= '3.8'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
pillow==10.4.0; python_version >= '3.8'
psycopg[binary,pool]==3.2.2; python_version >= '3.8'
psycopg-binary==3.2.2
psycopg-pool==3.2.3
pyjwt==2.9.0; python_version >= '3.8'
sqlparse==0.5.1; python_version >= '3.8'
typing-extensions==4.12.2; python_version < '3.13'
tzdata==2024.1; sys_platform == 'win32'
